+++ src/transformation/to_trimesh/convex_polyhedron_to_trimesh.rs
@@ -9,9 +9,9 @@ impl ConvexPolyhedron {
         for face in self.faces() {
             let i1 = face.first_vertex_or_edge;
             let i2 = i1 + face.num_vertices_or_edges;
-            let first_id = self.vertices_adj_to_face()[i1] as u32;
+            let first_id = self.vertices_adj_to_face()[i1 as usize] as u32;
 
-            for idx in self.vertices_adj_to_face()[i1 + 1..i2].windows(2) {
+            for idx in self.vertices_adj_to_face()[i1 as usize + 1..i2 as usize].windows(2) {
                 indices.push(Point3::new(first_id, idx[0] as u32, idx[1] as u32));
             }
         }
